+++ nucypher/blockchain/eth/networks.py
@@ -21,6 +21,8 @@ class NetworksInventory:  # TODO: See #1564
     MAINNET = 'mainnet'
     IBEX = 'ibex'
     LYNX = 'lynx'
+    ORYX = 'oryx'
+    ETH = 'ethereum'
 
     # TODO: Use naming scheme to preserve multiple compatibility with multiple deployments to a single network?
     POLYGON = 'polygon'
@@ -29,17 +31,22 @@ class NetworksInventory:  # TODO: See #1564
     UNKNOWN = 'unknown'  # TODO: Is there a better way to signal an unknown network?
     DEFAULT = MAINNET
 
-    __to_ethereum_chain_id = {
+    __to_chain_id_eth = {
         MAINNET: 1,      # Ethereum Mainnet
+        ORYX: 3,         # Ropsten
         IBEX: 4,         # Rinkeby
         LYNX: 5,         # Goerli
-
+    }
+    __to_chain_id_polygon = {
         # TODO: Use naming scheme?
         POLYGON: 137,    # Polygon Mainnet
         MUMBAI: 80001,   # Polygon Testnet (Mumbai)
     }
 
-    NETWORKS = tuple(__to_ethereum_chain_id.keys())
+    ETH_NETWORKS = tuple(__to_chain_id_eth.keys())
+    POLY_NETWORKS = tuple(__to_chain_id_polygon.keys())
+
+    NETWORKS = ETH_NETWORKS + POLY_NETWORKS
 
     class UnrecognizedNetwork(RuntimeError):
         pass