+++ nucypher/cli/actions/select.py
@@ -136,13 +136,23 @@ def select_client_account(emitter,
     return chosen_account
 
 
-def select_network(emitter: StdoutEmitter, message: Optional[str] = None) -> str:
+def select_network(emitter: StdoutEmitter, network_type: str, message: Optional[str] = None) -> str:
     """Interactively select a network from nucypher networks inventory list"""
-    emitter.message(message=message or str(), color='yellow')
-    rows = [[n] for n in NetworksInventory.NETWORKS]
-    emitter.echo(tabulate(rows, showindex='always'))
-    choice = click.prompt(SELECT_NETWORK, default=0, type=click.IntRange(0, len(NetworksInventory.NETWORKS)-1))
-    network = NetworksInventory.NETWORKS[choice]
+    emitter.message(message=message or str(), color="yellow")
+    if network_type == NetworksInventory.ETH:
+        network_list = NetworksInventory.ETH_NETWORKS
+    elif network_type == NetworksInventory.POLYGON:
+        network_list = NetworksInventory.POLY_NETWORKS
+    else:
+        raise(ValueError("Network type must be either 'eth' or 'polygon'"))
+    rows = [[n] for n in network_list]
+    emitter.echo(tabulate(rows, showindex="always"))
+    choice = click.prompt(
+        SELECT_NETWORK,
+        default=0,
+        type=click.IntRange(0, len(rows) - 1),
+    )
+    network = network_list[choice]
     return network