+++ nucypher/cli/commands/status.py
@@ -23,8 +23,8 @@ import click
 
 from nucypher.blockchain.eth.agents import (
     ContractAgency,
+    NucypherTokenAgent,
     PREApplicationAgent,
-    SubscriptionManagerAgent,
     EthereumContractAgent
 )
 from nucypher.blockchain.eth.constants import AVERAGE_BLOCK_TIME_IN_SECONDS
@@ -57,7 +57,7 @@ POLICY_MANAGER = 'PolicyManager'
 
 CONTRACT_NAMES = [
     PREApplicationAgent.contract_name,
-    SubscriptionManagerAgent.contract_name,
+    NucypherTokenAgent.contract_name,
     STAKING_ESCROW,
     POLICY_MANAGER
 ]
@@ -147,7 +147,7 @@ def staking_providers(general_config, registry_options, staking_provider_address
 @status.command()
 @group_registry_options
 @group_general_config
-@option_contract_name(required=False)
+@option_contract_name(required=False, valid_options=CONTRACT_NAMES)
 @option_event_name
 @option_from_block
 @option_to_block
@@ -161,7 +161,7 @@ def events(general_config, registry_options, contract_name, from_block, to_block
     if csv or csv_file:
         if csv and csv_file:
             raise click.BadOptionUsage(option_name='--event-filter',
-                                       message=f'Pass either --csv or --csv-file, not both.')
+                                       message=click.style('Pass either --csv or --csv-file, not both.', fg="red"))
 
         # ensure that event name is specified - different events would have different columns in the csv file
         if csv_file and not all((event_name, contract_name)):
@@ -169,12 +169,14 @@ def events(general_config, registry_options, contract_name, from_block, to_block
             #  - each appended event adds their column names first
             #  - single report-type functionality, see #2561
             raise click.BadOptionUsage(option_name='--csv-file, --event-name, --contract_name',
-                                       message='--event-name and --contract-name must be specified when outputting to '
-                                               'specific file using --csv-file; alternatively use --csv')
+                                       message=click.style('--event-name and --contract-name must be specified when outputting to '
+                                               'specific file using --csv-file; alternatively use --csv', fg="red"))
     if not contract_name:
         if event_name:
-            raise click.BadOptionUsage(option_name='--event-name', message='--event-name requires --contract-name')
+            raise click.BadOptionUsage(option_name='--event-name', message=click.style('--event-name requires --contract-name', fg="red"))
         # FIXME should we force a contract name to be specified?
+        # default to PREApplication contract
+        contract_names = [PREApplicationAgent.contract_name]
     else:
         contract_names = [contract_name]
 
@@ -190,8 +192,8 @@ def events(general_config, registry_options, contract_name, from_block, to_block
         # validate block range
         if from_block > to_block:
             raise click.BadOptionUsage(option_name='--to-block, --from-block',
-                                       message=f'Invalid block range provided, '
-                                               f'from-block ({from_block}) > to-block ({to_block})')
+                                       message=click.style(f'Invalid block range provided, '
+                                               f'from-block ({from_block}) > to-block ({to_block})', fg="red"))
 
     # event argument filters
     argument_filters = None
@@ -200,8 +202,8 @@ def events(general_config, registry_options, contract_name, from_block, to_block
             argument_filters = parse_event_filters_into_argument_filters(event_filters)
         except ValueError as e:
             raise click.BadOptionUsage(option_name='--event-filter',
-                                       message=f'Event filter must be specified as name-value pairs of '
-                                               f'the form `<name>=<value>` - {str(e)}')
+                                       message=click.style(f'Event filter must be specified as name-value pairs of '
+                                               f'the form `<name>=<value>` - {str(e)}', fg="red"))
 
     emitter.echo(f"Retrieving events from block {from_block} to {to_block}")
 
@@ -209,7 +211,7 @@ def events(general_config, registry_options, contract_name, from_block, to_block
     if legacy and contract_name in LEGACY_CONTRACT_VERSIONS:
         contract_version = LEGACY_CONTRACT_VERSIONS[contract_name]
 
-    for contract_name in CONTRACT_NAMES:
+    for contract_name in contract_names:
         if legacy:
             versioned_contract = blockchain.get_contract_by_name(
                 registry=registry,
@@ -229,7 +231,7 @@ def events(general_config, registry_options, contract_name, from_block, to_block
 
         if event_name and event_name not in agent.events.names:
             raise click.BadOptionUsage(option_name='--event-name, --contract_name',
-                                       message=f'{contract_name} contract does not have an event named {event_name}')
+                                       message=click.style(f'{contract_name} contract does not have an event named {event_name}', fg="red"))
 
         title = f" {agent.contract_name} Events ".center(40, "-")
         emitter.echo(f"\n{title}\n", bold=True, color='green')