+++ tests/constants.py
@@ -100,6 +100,8 @@ PYEVM_DEV_URI = "tester://pyevm"
 
 TEST_ETH_PROVIDER_URI = PYEVM_DEV_URI  # TODO: Pytest flag entry point?
 
+TEST_POLYGON_PROVIDER_URI = "tester://polygon"
+
 MOCK_ETH_PROVIDER_URI = 'tester://mock'
 
 #