+++ tests/fixtures.py
@@ -458,6 +458,7 @@ def federated_porter(federated_ursulas):
                     known_nodes=federated_ursulas,
                     verify_node_bonding=False,
                     federated_only=True,
+                    execution_timeout=2,
                     network_middleware=MockRestMiddleware())
     yield porter
     porter.stop_learning_loop()
@@ -471,6 +472,7 @@ def blockchain_porter(blockchain_ursulas, testerchain, test_registry):
                     known_nodes=blockchain_ursulas,
                     eth_provider_uri=TEST_ETH_PROVIDER_URI,
                     registry=test_registry,
+                    execution_timeout=2,
                     network_middleware=MockRestMiddleware())
     yield porter
     porter.stop_learning_loop()
@@ -501,7 +503,7 @@ def _make_testerchain(mock_backend: bool = False) -> TesterBlockchain:
     import eth
     eth._utils.headers.GAS_LIMIT_MINIMUM = TEST_GAS_LIMIT
     eth._utils.headers.GENESIS_GAS_LIMIT = TEST_GAS_LIMIT
-    eth.vm.forks.frontier.headers.GENESIS_GAS_LIMIT = TEST_GAS_LIMIT
+    eth.vm.forks.london.headers.GENESIS_GAS_LIMIT = TEST_GAS_LIMIT
 
     # Monkey patch to prevent gas estimates
     def _get_buffered_gas_estimate(web3, transaction, gas_buffer=100000):